Revert "SystemUI: Enable fading squiggle animation of qs media player"
This reverts commit 07f5564056880e1cc993dd34fca0f803a9885236.

Reason for revert: The squiggly line exists as an indicator that that
part of the song has already been played. The rest of the song, that
hasn’t been played is expressed as a flat line. The knob indicates
where the song is currently at.

With the above stated, it would make sense not to extend the...
squiggliness beyond the knob’s position, as that part of the song has
not been reached yet.

(Talking from purely a design standpoint)

Change-Id: Ic672d8404b2e2613dbc114b3c528184c3844a828